stable version 1
Is good enough to run the Fibonacci sequence and a for Loop. There is a bug where: `#var a,b;{#set a=b;/*creates a new variable for a*/}`.

adder works

i think this one is good enough to run the Fibonacci sequence and a for Loop. There is a bug where: `#var a,b;{#set a=b;/*creates a new variable for a*/}`.